requirements.txt
transformers
sentence-transformers
faiss-cpu
torch
flask
numpy